Support Series plugin
The Pelican Series plugin (https://github.com/pelican-plugins/series)
is a convenient way for authors to link articles together as multiple
installments of a series.

Add support for the Series plugin to this theme: render articles
belonging to a series with a footer pointing to the preceding and
following articles in the series, using the same styling as "related
posts" links.

This implementation is heavily inspired by the Series plugin in the
Pelican bootstrap3 theme (MIT licensed):
https://github.com/getpelican/pelican-themes/blob/master/pelican-bootstrap3/templates/includes/series.html